# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Task:
    properties:
      create_by:
        type: integer
      created_at:
        type: string
      describe:
        type: string
      finished_at:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      require:
        type: string
      status:
        type: string
    type: object
  entity.TaskRun:
    properties:
      created_at:
        type: string
      duration:
        type: integer
      end_at:
        type: string
      id:
        type: integer
      start_at:
        type: string
      status:
        type: string
      task_id:
        type: integer
    type: object
  entity.TaskRunLog:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      images:
        type: string
      task_id:
        type: integer
      task_run_id:
        type: integer
      user_id:
        type: integer
      videos:
        type: string
    type: object
...
  http_util.Response:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/app_code.CodeType'
        example: success
      data: {}
      error:
        example: error message
        type: string
    type: object
  v1.Task:
    properties:
      describe:
        type: string
      location:
        type: string
      name:
        type: string
      require:
        type: string
    type: object
...
host: task.mytool.zone
info:
  contact: {}
  description: Task Mini Program API
  title: Task Mini Program
  version: "1.0"
paths:
...
  /task/apply:
    post:
      consumes:
      - application/json
      description: 用户申请参加任务
      operationId: apply-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Apply task
      tags:
      - task
  /task/assignRole:
    post:
      consumes:
      - application/json
      description: 管理员分配任务角色
      operationId: assign-role
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      - description: userID
        in: query
        name: userID
        required: true
        type: integer
      - description: role
        enum:
        - leader
        - member
        - recorder
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Assign role
      tags:
      - task
  /task/auditApplyTask:
    post:
      consumes:
      - application/json
      description: 管理员审核用户参加任务
      operationId: audit-user-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      - description: userID
        in: query
        name: userID
        required: true
        type: integer
      - description: auditStatus
        enum:
        - rejected
        - approved
        in: query
        name: auditStatus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Audit user task
      tags:
      - task
  /task/auditTask:
    post:
      consumes:
      - application/json
      description: 管理员审核任务
      operationId: audit-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      - description: auditStatus
        enum:
        - rejected
        - approved
        in: query
        name: auditStatus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Audit task
      tags:
      - task
  /task/cancel:
    post:
      consumes:
      - application/json
      description: 取消任务
      operationId: cancel-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Cancel task
      tags:
      - task
  /task/create:
    post:
      consumes:
      - application/json
      description: 创建任务
      operationId: create-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建任务
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/v1.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Create task
      tags:
      - task
  /task/detail:
    get:
      consumes:
      - application/json
      description: 获取任务详情
      operationId: detail
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Show task detail
      tags:
      - task
  /task/finish:
    post:
      consumes:
      - application/json
      description: 完成任务
      operationId: finish-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Finish task
      tags:
      - task
  /task/list:
    get:
      consumes:
      - application/json
      description: 获取任务列表
      operationId: list
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: lastID
        in: query
        name: lastID
        type: integer
      - description: status
        enum:
        - torun
        - audit_fail
        - join
        - new
        - running
        - paused
        - finished
        - canceled
        - deleted
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Task'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: List tasks
      tags:
      - task
  /task/pause:
    post:
      consumes:
      - application/json
      description: 暂停任务
      operationId: pause-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Pause task
      tags:
      - task
  /task/prepare:
    get:
      consumes:
      - application/json
      description: 获取签到二维码
      operationId: prepare-task
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.TaskRun'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Prepare task
      tags:
      - task
  /task/runLogs:
    get:
      consumes:
      - application/json
      description: 获取任务运行日志列表
      operationId: run-log-list
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: taskID
        in: query
        name: taskID
        required: true
        type: integer
      - description: lastID
        in: query
        name: lastID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.TaskRunLog'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: Run log list
      tags:
      - task
...
swagger: "2.0"